# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlixQuest Scraper API",
    "version": "2.0.0",
    "description": "Exhaustive OpenAPI specification for FlixQuest Scraper API v2, designed for integration with Flutter applications."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/v2/dlhd/channels": {
      "get": {
        "summary": "List DLHD Live Channels",
        "description": "Scrapes the current 24/7 channel directory from DLHD.",
        "operationId": "getDlhdChannels",
        "tags": ["DLHD"],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Case-insensitive channel name or ID filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "description": "Set to true to bypass the channel cache.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current DLHD channels.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "source", "count", "channels"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "source": {
                      "type": "string",
                      "example": "dlhd"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "channels": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DlhdChannel"
                      }
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "DLHD channel directory was unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/dlhd/channels/{id}/stream": {
      "get": {
        "summary": "Extract a DLHD Channel HLS Stream",
        "description": "Resolves a numeric DLHD channel ID to a validated, short-lived M3U8 URL and the Origin, Referer, and User-Agent headers required for playback.",
        "operationId": "getDlhdChannelStream",
        "tags": ["DLHD"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Numeric DLHD channel ID.",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "example": "51"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fresh HLS URL and playback headers. The response is not cacheable.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "source",
                    "channel",
                    "stream",
                    "extractedAt"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "source": {
                      "type": "string",
                      "example": "dlhd"
                    },
                    "channel": {
                      "$ref": "#/components/schemas/DlhdChannel"
                    },
                    "stream": {
                      "$ref": "#/components/schemas/DlhdStream"
                    },
                    "extractedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The channel ID is not numeric.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "No current playable M3U8 could be extracted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/dlhd/epg": {
      "get": {
        "summary": "Get the Categorized DLHD EPG",
        "description": "Returns the DLHD schedule grouped by day, category, event, and playable channels.",
        "operationId": "getDlhdEpg",
        "tags": ["DLHD"],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "description": "Schedule date in YYYY-MM-DD format.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "description": "Case-insensitive category filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Search event titles, times, and channel names.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "description": "Set to true to bypass the EPG cache.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categorized DLHD EPG.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DlhdEpgResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid date format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "DLHD EPG was unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DlhdChannel": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": {
            "type": "string",
            "example": "51"
          },
          "name": {
            "type": "string",
            "example": "ABC USA"
          },
          "letter": {
            "type": "string",
            "example": "A"
          },
          "watchUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://dlhd.st/watch.php?id=51"
          },
          "playerUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://dlhd.st/stream/stream-51.php"
          }
        }
      },
      "DlhdStream": {
        "type": "object",
        "required": ["url", "isM3U8", "headers", "embedUrl"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "A validated, short-lived HLS master playlist URL."
          },
          "isM3U8": {
            "type": "boolean",
            "enum": [true]
          },
          "headers": {
            "type": "object",
            "description": "Headers required for the HLS master, variants, and segments.",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "Accept": "*/*",
              "Origin": "https://player-host.example",
              "Referer": "https://player-host.example/",
              "User-Agent": "Mozilla/5.0 ..."
            }
          },
          "embedUrl": {
            "type": "string",
            "format": "uri"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DlhdEpgResponse": {
        "type": "object",
        "required": [
          "success",
          "source",
          "timezone",
          "dayCount",
          "eventCount",
          "days"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "source": {
            "type": "string",
            "example": "dlhd"
          },
          "timezone": {
            "type": "string",
            "example": "UK GMT"
          },
          "filters": {
            "type": "object",
            "additionalProperties": true
          },
          "dayCount": {
            "type": "integer"
          },
          "eventCount": {
            "type": "integer"
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DlhdEpgDay"
            }
          }
        }
      },
      "DlhdEpgDay": {
        "type": "object",
        "required": ["label", "categories"],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "label": {
            "type": "string"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DlhdEpgCategory"
            }
          }
        }
      },
      "DlhdEpgCategory": {
        "type": "object",
        "required": ["name", "slug", "events"],
        "properties": {
          "name": {
            "type": "string",
            "example": "PPV Events"
          },
          "slug": {
            "type": "string",
            "example": "ppv-events"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DlhdEpgEvent"
            }
          }
        }
      },
      "DlhdEpgEvent": {
        "type": "object",
        "required": ["id", "time", "title", "channels"],
        "properties": {
          "id": {
            "type": "string"
          },
          "time": {
            "type": "string",
            "example": "16:00"
          },
          "title": {
            "type": "string",
            "example": "APFC Indian Fight League"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "channels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DlhdChannel"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Missing or invalid tmdbId parameter"
          },
          "details": {
            "type": "string",
            "example": "Available providers: vixsrc, vidsrc, notorrent, ..."
          }
        }
      }
    }
  }
}